@@ -1,40 +1,66 @@
 # This file is used to configure your project.
 # Read more about the various options under:
-# http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
+# https://setuptools.pypa.io/en/latest/userguide/declarative_config.html
+# https://setuptools.pypa.io/en/latest/references/keywords.html
 
 [metadata]
 name = shimoku-api-python
 description = Shimoku enables you to build Data Products in just hours and allows you to create Predictive Analytics Products with Artificial Intelligence capabilities.
 author = Shimoku
-author-email = contact@shimoku.com
-license = mit
-long-description = file: README.md
-long-description-content-type = text/x-rst; charset=UTF-8
+author_email = contact@shimoku.com
+license = MIT
+license_files = LICENSE.txt
+long_description = file: README.rst
+long_description_content_type = text/x-rst; charset=UTF-8
 url = https://github.com/shimoku-tech/shimoku-api-python
-project-urls =
+project_urls =
     Documentation = https://docs.shimoku.com/
 # Change if running only on Windows, Mac or Linux (comma-separated)
 platforms = any
+
 # Add here all kinds of additional classifiers as defined under
-# https://pypi.python.org/pypi?%3Aaction=list_classifiers
+# https://pypi.org/classifiers/
 classifiers =
     Development Status :: 4 - Beta
     Programming Language :: Python
 
+
 [options]
 zip_safe = False
-packages = find:
+packages = find_namespace:
 include_package_data = True
 package_dir =
     =src
-# DON'T CHANGE THE FOLLOWING LINE! IT WILL BE UPDATED BY PYSCAFFOLD!
-setup_requires = pyscaffold>=3.2a0,<3.3a0
-# Add here dependencies of your project (semicolon/line-separated), e.g.
-# install_requires = numpy; scipy
-# The usage of test_requires is discouraged, see `Dependency Management` docs
-# tests_require = pytest; pytest-cov
-# Require a specific Python version, e.g. Python 2.7 or >= 3.4
-# python_requires = >=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*
+
+# Require a min/specific Python version (comma-separated conditions)
+python_requires = >=3.8
+
+# Add here dependencies of your project (line-separated), e.g. requests>=2.2,<3.0.
+# Version specifiers like >=2.2,<3.0 avoid problems due to API changes in
+# new major versions. This works if the required packages follow Semantic Versioning.
+# For more information, check out https://semver.org/.
+install_requires =
+    importlib-metadata
+    pandas~=2.1.3
+    requests~=2.31.0
+    datetime~=4.9
+    json5~=0.9.10
+    tenacity~=8.1.0
+    shimoku-components-catalog>=0.2.1
+    tqdm~=4.64.1
+    psutil~=5.9.4
+    aiohttp~=3.8.3
+    ipython>=8.10
+    uvicorn~=0.23.2
+    graphene~=3.3
+    urllib3~=1.26.13
+    pydantic~=2.4.2
+    strawberry-graphql~=0.198.0
+    starlette_graphene3~=0.6.0
+    fastapi~=0.103.2
+    websockets~=11.0.3
+    python-dateutil~=2.8.2
+    uvloop~=0.19.0
 
 [options.packages.find]
 where = src
